Add and wire up the signing interfaces.
This defines an interface to convert a TaskRun to a payload format. Right
now the only supported format is just native Tekton (a passthrough of the TaskRun).
We can add other formats, like InToto and Grafeas next.

This also adds a few other small functions so we can wire this up to the controller:
- MarkSigned, to mark the TaskRun as signed.
- IsSigned, to tell if the TaskRun has already been signed.

Note: Payloads are not currently stored anywhere.