Actually sign payloads!
This commit hooks up the secret and our PGP signing library to the reconciler.
It also tweaks the storage backend interface a bit - we need to take two parameters instead of one:
- The signed body (in byte form)
- The signature itself

Since we need to store both of these.

I also added some extra debugging to the e2e tests.