Migrate chains to use genreconciler.
This leverages genereconciler to dramatically reduce the boilerplate needed for tekton/chains and dramatically increase the code coverage.

Much of the churn in vendor is from `go mod tidy && go mod vendor`, which I ran to pull in a few previously pruned packages from `knative.dev/pkg`, and it seems to have drifted a bit.