Migrate the ConfigStore to use standard infrastructure.
Knative has a ton of utilities for managing config storage, watching configmaps, and snapshotting those configmaps onto context during each reconciliation.

This is really just a first chunk, but migrates the bulk of the config package to support the standard interfaces, use the libraries for validation, and leverage the shared configmap watching libraries.

In a subsequent change, it would make the most sense to have the signer extract the config store from the context it is passed instead of having its own reference, but this change has grown overly large already.

/kind cleanup