Access Config and Logger from context.
This changes the signer to access the config and the logger off of context instead of squirreling them away at construction.

For the configstore, this is about consistency.  The reconciliation framework snapshots configurations onto context at the beginning of reconciliation so that the configuration doesn't change in the middle of reconciliation.  By having the signer access the same config as the rest of the reconciler, we ensure that we don't get inconsistent views of the configuraton data.

For the logger, this is about tagging.  The reconciliation framework infuses the base logger with a variety of additional tagging, like the reconciliation key, which the current signer's logger doesn't have, so by switching to the logger that comes from the reconciliation context, we will get significantly richer log tagging from within the signer.