Change the ServiceAccount identity used for OCI storage backend.
Rather than hardcoding the chains controller as the service account identity used for OCI-based signing, use the task's service account identity.

This should just simplify things because users are already configuring the Task with credentials to push the image, but currently need to grant access to the chains controller as well.

Fixes: https://github.com/tektoncd/chains/issues/161