Drop GCS storage backend from unit testing.
The GCS storage backend checks auth when it is initialized, which only works if GCP auth is available in the execution environment (e.g. GCE, GCB), and fails if not (e.g. locally, github actions).

I'm attempting to add a "downstream test" to knative/pkg which runs the downstream unit tests on actions after updating PKG and codegen, but with this test things currently fail: https://github.com/knative/pkg/pull/2237

Fixes: https://github.com/tektoncd/chains/issues/211