Add `chains-info` ConfigMap with version info
This commit adds a `chains-info` ConfigMap which contains the version of
Chains installed on the cluster. This is meant to be primarily read by
other tools like `tkn` to display component version.

This ConfigMap can be read by someone who does not even have permissions
to view other resources in the `tekton-chains` namespace, hence making
it easier to advertise the version instead of having tools wrangle with
labels on deployments to which permissions might be limited.

This is in line with similar `<component>-info` ConfigMaps used in other
Tekton components like Pipeline and Triggers.

Fix #336