Add support for PubSub storage backend
This patch implements support for "in-memory" and "kafka" pubsub.

The in-memory pubsub is mopstly used for testing. The gcloud library
offers support for other pubsub provider, therefore this backend was
designed with this idea in mind and allow for easy implementation of
additional providers if needed in the future.

The experimental documentation was updated to explain how to enable this
feature.