Do not fail OCI upload when no subjects
Instead, log an INFO message and carry on. Many tasks in a given Tekton
Pipeline may not produce an image. For example, a Pipeline may contain
tasks to clone a git repository and run different kinds of tests.

Triggering an error means the TaskRun resource is re-added to the
operator reconcilliation loop, causing the same resource to be
re-processed multiple times. If other storage backends are also
configured, payloads are regenerated and re-uploaded to those backends.
This multiple upload behavior is also seen when rekor integration is
enabled.

Fixes #402

Signed-off-by: Luiz Carvalho <lucarval@redhat.com>