Put signers initialization back to SignTaskRun
Originally we created signers in configmap observer function same as
formatters and backends, but there is a potential issue with this.

For example, if x509 is selected initially, but later clients change
the signing secret, the signer will not get updated because the configmap is not changed.

So this commit, the signers creation is put back to SignTaskRun function.