Allow TaskRun payloads other than IMAGE_URL
Prior to this commit, in the `tekton` backend, a payload or a signature
was considered valid only if the TaskRun status had `IMAGE_URL` in the
results. This is incorrect since Chains considers both, `IMAGE_URL` and
`IMAGE_DIGEST` as results of interest. Moreover, Chains considers
PipelineResources to indicate an output of interest as well.

This commit removes checking for `IMAGE_URL` in the TaskRun status to
validate payloads, instead relies solely on the presence of a payload
annotation on the TaskRun given that it's only added when Chains has
signed something (after validation anyway).