Add more supported GCP KMS key ref formats in config doc
Prior to this commit, the config doc only mentions that the key ref
without version specified is supported. However, the key refs with
version specified are also specified. And there are two ways to specify
the version `../cryptoKeyVersions/..` (new) or `../versions/..` (old).

In this commit, we mention all support formats explicitly in the doc.