Fix param value ordering in materials
Prior to the fix, when generating `materials` section of the intoto
attestation, Chains gets the param value from TaskRun spec first and then
default from Task spec. This overwriting will cause the problem that
the actual value used in run is from taskrun spec, but the value recorded
in `materials` is the value defined in task spec's `default`.

In this PR, we fix the issue.