Use event updates for PR reconciliation (#564)
* Use event updates for PR reconciliation

A PipelineRun should only be processed when the underlying TaskRun tasks
have been signed. This prevents timing issues where the Chains
controller may attempt to push both TaskRun and PipelineRun attestations
to an OCI repository at the same time, causing one of them to be lost.

Prior to this change, a PipelineRun that was not ready to be processed
was requeued so it could be re-processed in 15 seconds. Given that this
scenario happens often, this caused the Chains controller to be
overwhelmed if there were too many TaskRun and/or PipelineRun resources
on the cluster.

This commit modifies the logic so when Chains signs a TaskRun, it also
updates an annotation, `chains.tekton.dev/child-signed`, on the parent
PipelineRun, if any. The annotation update triggers the PipelineRun to
be reconciled again. Thus the logic is driven by "events" instead of a
polling mechanism.

Signed-off-by: Luiz Carvalho <lucarval@redhat.com>

Notify parent object after child is reconciled

Signed-off-by: Luiz Carvalho <lucarval@redhat.com>

* Use informer to query for TaskRuns

Signed-off-by: Luiz Carvalho <lucarval@redhat.com>

* Use tracker to be notified of TaskRun updates

Signed-off-by: Luiz Carvalho <lucarval@redhat.com>

Signed-off-by: Luiz Carvalho <lucarval@redhat.com>