Add releases.md (#593)
Add a releases.md, in line with the other Tekton projects and
community guidelines.

Since chains does not have a fixed release schedule today, to be
inline with the community policy I included the 4 LTS releases and
any further release is marked as on demand.

To be aligned with other project, I included 1 month support for
non-LTS releases.

Fixes: #535

Signed-off-by: Andrea Frittoli <andrea.frittoli@gmail.com>

Signed-off-by: Andrea Frittoli <andrea.frittoli@gmail.com>