Convert times to UTC
The SLSA provenance prescribes that the Timestamp data types must be in
UTC time zone. It can happen that the environment where Chains
controller is running is not set to the UTC time zone, and in that case
the `predicate.metadata.buildStartedOn` and
`predicate.metadata.buildFinishedOn` will be set with the local time
zone. This is due to apimachinery converting `time.Time` values into
local timezone on JSON unmarshall.

This forces all dates to the UTC time zone.

Fixes #630