Generalize format loading. (#603)
* Generalize format loading.

This refactors how formatters are initialized and configured to be more
uniform. This allows for formatter providers to be more easily slotted
in / configured without needing to make modifications for downstream
signers. The ultimate goal for this change is to allow for "type
aliasing" of formatters, where the same formatter may be referenced by
different names to allow for supporting new formats in a backwards
compatible way (i.e. letting slsa be an alias for intoto or slsa0.2,
etc.)

This follows a similar pattern to sigstore KMS providers - https://github.com/sigstore/sigstore/tree/main/pkg/signature/kms

As part of this change, the way to initialize a formatter is
standardized. To support this, the intoto logger that was previously
configured during initialization is now taken from the context.

BREAKING CHANGES:
- formats.PayloadType is now defined in config rather than formats. The
  semantics behind this value are still the same.
- CreatePayload now takes in a context.
- InTotoIte6.NewFormatter no longer takes in a logger.
- formats.AllFormatters is removed. Use formats.GetPayloader instead.

Signed-off-by: Billy Lynch <billy@chainguard.dev>

* Add formatter loaders to controllers.

Signed-off-by: Billy Lynch <billy@chainguard.dev>