Cleanup Materials usage
Prior to this creation of materials utilized bad go-practice.
It seems bad practice to pass a slice pointer around. Other functions on the materials package deal with adding things to materials but don't take a pre-existing slice, e.g. `AddMaterialsFromTaskParamsAndResults`.

This PR attempts to clean it up and address issue
https://github.com/tektoncd/chains/issues/816.