Make child TaskRuns accessible
Prior, the child TaskRuns of a PipelineRunObject are not accessible directly.
Only way to access it is to loop over task name from pipeline spec and use
GetTaskRunFromTask method to get the taskrun object.

Now, we introduce a method to expose the slice of child taskruns directly so that
we can easily access all child taskruns, though the old GetTaskRunFromTask
is still useful in one case i.e. 0.2 buildConfig for which we do need to loop
over PipelineTask type and call it to get taskrun.

Signed-off-by: Chuang Wang <chuangw@google.com>