TektonObject specifies supported artifacts
Change the code to no longer require callers to inspect the underlying
type of the TektonObject. Instead, add a set of methods that specify
which artifacts are supported.

An alternative approach was consider to introduce a method such as
`TektonObject.GetSignables` to return `[]artifacts.Signable`. However,
`Signable` already has a dependency on `TektonObject` causing a circular
import.

Signed-off-by: Luiz Carvalho <lucarval@redhat.com>