Increase usage of TektonObject (#881)
* Add TektonObject.GetKindName to abstract more

This commit removes certain parts of the code that switch on the
underlying TektonObject type to determine if either the string `taskrun`
or `pipelinerun` should be used. Instead, a new method on TektonObject
is used.

Signed-off-by: Luiz Carvalho <lucarval@redhat.com>

* TektonObject specifies supported artifacts

Change the code to no longer require callers to inspect the underlying
type of the TektonObject. Instead, add a set of methods that specify
which artifacts are supported.

An alternative approach was consider to introduce a method such as
`TektonObject.GetSignables` to return `[]artifacts.Signable`. However,
`Signable` already has a dependency on `TektonObject` causing a circular
import.

Signed-off-by: Luiz Carvalho <lucarval@redhat.com>

* Increase test coverage of TektonObjects

Signed-off-by: Luiz Carvalho <lucarval@redhat.com>

---------

Signed-off-by: Luiz Carvalho <lucarval@redhat.com>