Manage consts of ResourceDescriptor in its own pkg
Prior, there are a number of consts defined for the ResourceDescriptor
across different packages, which introduces dependencies between those
packages i.e. taskrun pkg imports pipelinerun pkg just for json media type,
which in turn is prone to import cycle in future.

Now, since ResourceDescriptor in SLSA v1 is a standalone type, we
organize and manage all the constants for the fields of ResourceDescriptor
into its own (internal) package.

Signed-off-by: Chuang Wang <chuangw@google.com>