Add support for SBOM attestations
This commit teaches Chains to create SBOM attestations based on type
hinting results from TaskRuns.

The general idea is that a Task generates the SBOM, then uploads it as a
blob to an OCI registry. Chains then downloads the blob and uses it as
the payload when creating the SBOM in-toto attestation.

Multiple SBOMs per image are allowed. They may be of different formats.

Signed-off-by: Luiz Carvalho <lucarval@redhat.com>