converting a few info into debug message
A taskRun/pipelineRun is watched as soon as created. Until the run is complete
and ready to proceed further, the chains controller is logging that the
taskRun/pipelineRun is still running at the info level. This looks good
but generates a lot of info in the controller logs and exponentially
increases with the number of taskRun and pipelineRun. Converting such message
to be reported at the debug log level.

Formatting a few error message such that the namespace/name is enclosed with
parenthesis so that its easy to read.

Signed-off-by: Priti Desai <pdesai@us.ibm.com>