Adds initial Attestor implementation.
This is the initial implementation of Attestors, which uses generics to
link chains components together with strict typing.

To start, this adds Attestor implementations of OCI signing and v1 SLSA
attestations. These Attestors are NOT wired up to the controller yet,
since they don't yet support the full range of config options (and
there's likely a few tweaks we need to make in order to help reuse
components like signers between Attestors).