Expose performance options
This allow admins to specify a few parameters to better suit their use
of Chains.

`--threads-per-controller` controls the number of concurrent threads the
Chains controller processes. The default value is 2.

`--kube-api-burst` controle the maximum burst for throttle.

`--kube-api-qps` controles the maximum QPS to the server from the
client.

The approach taken here is the same one used by the Tekton Pipeline
controller for the sake of consistency in the ecosystem.

Signed-off-by: Luiz Carvalho <lucarval@redhat.com>