Expose performance options (v0.19) (#1081)
* Expose performance options

This allow admins to specify a few parameters to better suit their use
of Chains.

`--threads-per-controller` controls the number of concurrent threads the
Chains controller processes. The default value is 2.

`--kube-api-burst` controle the maximum burst for throttle.

`--kube-api-qps` controles the maximum QPS to the server from the
client.

The approach taken here is the same one used by the Tekton Pipeline
controller for the sake of consistency in the ecosystem.

Signed-off-by: Luiz Carvalho <lucarval@redhat.com>

* Remove test for Pipelines v0.41

This release is no longer supported and is no longer compatible with the
e2e tests in this branch.

Signed-off-by: Luiz Carvalho <lucarval@redhat.com>

---------

Signed-off-by: Luiz Carvalho <lucarval@redhat.com>