golangci-lint: only flag new issues (#1105)
Intoto has deprecated the old Subject types in favor of a new library,
and this is getting pulled in via recent cosign updates.
This is going to be a significant refactor that we'll need to make to
chains. This changes the behavior of the lint tool to only flag/block
on new issues for now to give us time to work through this.