Handle signing and attesting blobs
Chains fetches existing information about the container images it signs
and attests. This works well when the image reference refers to an Image
Index or an Image Manifest since those are served from the same endpoint
in a container registry. However, when fetching information about a
blob (layer) a different endpoint is needed.

Cosign handles this behavior by making this step optional:
https://github.com/sigstore/cosign/blob/c86498055d0c4ea2f39076064aa094db12f85f6a/cmd/cosign/cli/sign/sign.go#L181-L186

Thus it is possible to sign/attest a blob with the cosign CLI. This
commit implements the same logic to Chains so it can also sign/attest
blobs.

Signed-off-by: Luiz Carvalho <lucarval@redhat.com>