Use server-side apply for managing annotations (#1411)
This is part of broader effort to switch all Tekton components to using
Server-Side Apply patch to improve performance (less retries when merge
patches are combined with resourceVersion) and reliability (when merge
patches are used without resourceVersion). The underlying problem is
that in the case of PipelineRun and TaskRun objects, there are multiple
controllers operating on those objects and that leads to issues such as
one controller overriding the changes (annotations, labels) of other
controllers.