Refactor the annotations management (#1422)
Annotations management moved to a separate package, so it can be
imported by other packages and avoid cycling import.
The tekton package now imports the annotations package and uses the
AddAnnotations function, while previously it duplicated the same
functionality.
The only functionlity previously provided by the patch package was the
CreateAnnotationsPatch function, which is now moved to the annotations
package and the patch package removed.
No new functionality added or removed.